requirements.txt
APScheduler==3.6.3
Django==2.2.7
django-cors-headers==3.10.0
django-mssql==1.8
django-mssql-backend==2.8.1
django-rest-swagger==2.2.0
django-sqlserver==1.11
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
pandas==1.0.3
Pillow==8.4.0
PyJWT==1.7.1
pyodbc==4.0.32
xlrd==1.2.0
XlsxWriter==3.0.2
xlwt==1.3.0
pymysql==1.0.2
pymssql==2.2.2
numpy==1.21.4
